feat: Added debounce, showing shelf on search page, added "books not found" and "couldn't search" texts to inform the user.
Added debounce to call the API less frequently after waiting a bit to 
give the user a chance to finish typing.
SearchPage should also show the shelf of the book but it wasn't because 
the API was not giving information about the books' shelf. This is fixed 
by comparing the data returned by API with the data we have from 
getAllBooks.
Added two informative messages in order to inform the user once search 
API failed or there was no book returned.